Now obj spawn at good position for everybody

+++ Assets/Script/Factory.cs
@@ -26,17 +26,17 @@ public class Factory : MonoBehaviourPunCallbacks
         }
     }
 
-    public void factoryGO(double tx, double ty, double tz, double rx, double ry, double rz, double ox, double oy, double oz, int playerID, int index)
+    public void factoryGO(float tx, float ty, float tz, float rx, float ry, float rz, float ox, float oy, float oz, int playerID, int index)
     {
         photonView.RPC("factoryGOpun", RpcTarget.All, tx,ty,tz,rx,ry,rz,ox,oy,oz,playerID,index);
     }
 
     [PunRPC]
-    public void factoryGOpun(double tx, double ty, double tz, double rx, double ry, double rz, double ox, double oy, double oz, int playerID, int index)
+    public void factoryGOpun(float tx, float ty, float tz, float rx, float ry, float rz, float ox, float oy, float oz, int playerID, int index)
     {
-        Vector3 pos = new Vector3((float)tx, (float)ty, (float)tz);
-        Vector3 rot = new Vector3((float)rx, (float)ry,(float)rz);
-        Vector3 offset = new Vector3((float)ox, (float)oy,(float)oz);
+        Vector3 pos = new Vector3(tx, ty, tz);
+        Vector3 rot = new Vector3(rx, ry,rz);
+        Vector3 offset = new Vector3(ox, oy,oz);
         GameObject go = PhotonNetwork.Instantiate(Objects[index].name, pos + offset, Quaternion.Euler(rot));
         go.AddComponent<Treasure>();
         go.GetComponent<Treasure>().playerID = playerID;
